requirements.txt
Faker==37.4.0
fpdf==1.7.2
jupyter_client==8.6.3
jupyter_core==5.8.1
numpy==2.2.6
openai==1.90.0
pandas==2.3.0
python-dotenv==1.1.0
tiktoken==0.4.0
tabulate==0.9.0